Create parser directory and move JSON parser to the directory
