+++ parser/generic_test.go
@@ -1,15 +1,16 @@
-package rlogs_test
+package parser_test
 
 import (
 	"testing"
 
 	"github.com/m-mizutani/rlogs"
+	"github.com/m-mizutani/rlogs/parser"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
 func TestJSONParserTimestamp(t *testing.T) {
-	psr := rlogs.JSONParser{
+	psr := parser.JSON{
 		TimestampField:  rlogs.String("ts"),
 		TimestampFormat: rlogs.String("2006-01-02T15:04:05"),
 	}
@@ -30,7 +31,7 @@ func TestJSONParserTimestamp(t *testing.T) {
 }
 
 func TestJSONParserUnixtime(t *testing.T) {
-	psr := rlogs.JSONParser{
+	psr := parser.JSON{
 		UnixtimeField:   rlogs.String("unix"), // prioritized
 		TimestampField:  rlogs.String("ts"),
 		TimestampFormat: rlogs.String("2006-01-02T15:04:05"),
@@ -47,7 +48,7 @@ func TestJSONParserUnixtime(t *testing.T) {
 }
 
 func TestJSONParserUnixtimeMilliSeconds(t *testing.T) {
-	psr := rlogs.JSONParser{
+	psr := parser.JSON{
 		UnixtimeMilliField: rlogs.String("unix"),
 	}
 	logs, err := psr.Parse(&rlogs.MessageQueue{
@@ -62,7 +63,7 @@ func TestJSONParserUnixtimeMilliSeconds(t *testing.T) {
 }
 
 func TestJSONParserNoTimestampField(t *testing.T) {
-	psr := rlogs.JSONParser{}
+	psr := parser.JSON{}
 	_, err := psr.Parse(&rlogs.MessageQueue{
 		Raw: []byte(`{"color":"blue","ts":"2019-10-19T04:44:44"}`),
 		Src: &rlogs.AwsS3LogSource{Region: "test-r", Bucket: "test-b", Key: "test-k"},
